Create OneWay.Py
This script takes in a csv, which I stuck in the same directory. I'll upload that csv for completeness. It's just a hypothetical example. Anyway, it takes that data and places it into a pandas dataframe. From there it's able to create a boxplot of the data, give us a thorough summary of the data, and run two normality tests on each column in the dataframe. In this case, there are three columns in the data frame. Ideally, we would confirm that our data is normal before running the final one-way ANOVA test. 

In this case, we'll see that our data is possibly normal. One test will fail for normality (Shapiro-Wilk), and one will pass (K^2). In the end, we end up having to reject the null hypothesis of the One-Way ANOVA. The nice thing is we can get an idea why it fails because of all the testing we do prior to running through the One-Way ANOVA.